--- API_model/a_deployEC2.md
@@ -1,73 +1,0 @@
-https://aws.plainenglish.io/deploy-docker-image-to-aws-ec2-in-5-minutes-4cd7518feacc
-
-
-## Docker
-
-* Build the image properly
-If you build your Docker image natively on an M1, Docker will create an ARM64 image by default.
-→ That image might not run properly on an AMD64-based server unless the target platform supports multi-arch images.
-
-docker buildx build --platform linux/amd64 -t <tag_image> .
-
-
-* Go to Docker Hub and create a new repository.
-
-* push your Docker image to Docker Hub. It’s preferable to push it to the Private repository rather than the Public repository
-
-## AWS
-
-* Create an EC2 Instance and install Docker on it
-
-1. Go to the EC2 dashboard in AWS Console and click the Launch instance button.
-Press enter or click to view image in full size
-
-2. Name your instance, for example: node-docker-server. The image should be Amazon Linux 2023. If you want a free tier-eligible instance, choose t2.micro.
-Press enter or click to view image in full size
-
-3. We would like to login to the instance, and we will create a key pair. Security group should have open inbound ports 22 and 3000.
-Press enter or click to view image in full size
-
-4. For storage, I will leave default settings, I don’t need more than 8 GiB of storage. Don’t forget: Free tier eligible customers can get up to 30 GB of EBS General Purpose (SSD) or Magnetic storage.
-Press enter or click to view image in full size
-
-5. Under Configure Storage, click the Advanced details.
-Press enter or click to view image in full size
-
-Go to the bottom of the page, and you will see the User data window. Paste this code into User data:
-"""
-#!/bin/bash
-sudo yum install docker -y
-sudo systemctl start docker
-sudo systemctl enable docker 
-"""
-Press enter or click to view image in full size
-From the right side, click the Launch instance button.
-
-## Deploy an image from Docker Hub to AWS EC2
-
-Deploy an image from Docker Hub to AWS EC2
-
-The first step is to connect to the EC2 instance. I will use AWS Console this time, but you can also connect through your Local machine from the terminal. Check out my older articles and you will find how to connect through a terminal from a Local machine.
-
-Switch to sudo mode with this command:
-
-sudo su
-
-Check the Docker with:
-
-docker --version
-
-Everything is set up, the next step is to deploy the image from Docker Hub. Use this command:
-
-docker run -d -p 3000:3000 <docker-hub-repo-name> 
-
-Press enter or click to view image in full size
-
-The image has been successfully created. Check the Public IP of your Instance with a 3000 port.
-Press enter or click to view image in full size
-
-We got our Docker application on the EC2 instance.
-
-    Let’s summarize: Thanks to AWS, we can create a server and deploy whatever we want to that server in just a few seconds. This is an example of deploying from Docker Hub to staging environment (EC2 instance) in a few seconds.
-
-
