+++ API_model/schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class PredictRequest(BaseModel):
+    data: list
+
+class LoginRequest(BaseModel):
+    username: str
+    password: str
