+++ API_model/troubleshout.md
@@ -1,3 +1,4 @@
 
 Start API
 uvicorn main:app --reload
+
