--- MODEL_train.py
@@ -1,9 +1,0 @@
-import numpy as np
-
-
-# Load files
-X_train_loaded = np.loadtxt("src/data/X_train.csv", delimiter=",")
-X_test_loaded = np.loadtxt("src/data/X_test.csv", delimiter=",")
-y_train_loaded = np.loadtxt("src/data/y_train.csv", delimiter=",")
-y_test_loaded = np.loadtxt("src/data/y_test.csv", delimiter=",")
-
